An incomplete rework for graphics component
I figured it's better to make graphics components to only hold data and
create a render system to handle the rendering, thus data rendered using
the same shader could be grouped togethre for better performace. So I'll
not try to fix all graphics components because they'll be refactored in
the future anyway.